feature-service/requirements.txt
# REST API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Message Queue
aio-pika>=9.0.0

# Database
asyncpg>=0.29.0

# HTTP Client
httpx>=0.25.0

# Configuration
pydantic-settings>=2.0.0
pyyaml>=6.0.0

# Logging
structlog>=23.2.0

# Time Series Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Parquet Storage
pyarrow>=14.0.0

# Orderbook Management
sortedcontainers>=2.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

